requirements.txt
PyYAML
torch>=1.9.0
numpy>=1.17
scipy>=1.2.1
prefetch_generator>=1.0.1
msgpack>=0.5.6
sortedcontainers
uvloop>=0.14.0
grpcio-tools>=1.33.2
protobuf>=3.12.2
configargparse>=1.2.3
py-multihash>=0.2.3
multiaddr @ git+https://github.com/multiformats/py-multiaddr.git@e01dbd38f2c0464c0f78b556691d655265018cce
cryptography>=3.4.6
pydantic>=2.0.0
packaging>=20.9
tenacity>=8.2.3
python-dotenv>=1.0.1
substrate-interface>=1.0.0